Move extension methods.
